quantum/requirements.txt
absl-py
numpy
cirq
matplotlib